建立 Android 14+ 平台基线
统一 Android 14+ 的 Gradle、SDK、AndroidX、Compose、ViewModel 和 Hilt 基线，
并保留单一 app 模块与现有应用身份。补齐 edge-to-edge、Photo Picker 受管文件、
上传 MIME、附件生命周期和 release 签名 fail-closed 约束。

验证：20 个 JVM 测试通过，lint 0 errors，debug 构建与配置缓存复用通过；
无签名 release 按预期失败，Trellis 安全与任务校验通过。

+++ app/src/main/java/org/houxg/leamonax/database/AppDataBase.java
@@ -100,7 +100,7 @@ public class AppDataBase {
                 int id = cursor.getInt(idIndex);
                 Account account = SQLite.select()
                         .from(Account.class)
-                        .where(Account_Table.id.eq(id))
+                        .where(Account_Table.id.eq((long) id))
                         .querySingle(database);
                 if (account != null) {
                     SQLite.update(Account.class)
@@ -144,7 +144,7 @@ public class AppDataBase {
                 int id = cursor.getInt(idIndex);
                 Account account = SQLite.select()
                         .from(Account.class)
-                        .where(Account_Table.id.eq(id))
+                        .where(Account_Table.id.eq((long) id))
                         .querySingle(database);
                 if (account != null) {
                     account.updateLastUseTime();
